Rename tests and function name.
