+++ src/services/comp-sys/component/tests/init-order/Components.toml
@@ -1,0 +1,5 @@
+# template
+[components]
+init-order = {name = "init-order"}
+first-init = {name = "first-init"}
+second-init = {name = "second-init"}
