+++ src/services/comp-sys/component/tests/init-order/tests/test.rs
@@ -1,17 +1,15 @@
-use bar::INIT_COUNT;
+use first_init::HAS_INIT;
 use component::init_component;
 use std::sync::atomic::Ordering::Relaxed;
 
 #[init_component]
 fn kernel_init() -> Result<(), component::ComponentInitError> {
-    assert_eq!(INIT_COUNT.load(Relaxed), 1);
-    INIT_COUNT.fetch_add(1, Relaxed);
     Ok(())
 }
 
 #[test]
 fn test() {
     simple_logger::init_with_level(log::Level::Debug).unwrap();
-    component::init(component::generate_information!()).unwrap();
-    assert_eq!(INIT_COUNT.load(Relaxed), 2);
+    component::init_all(component::parse_metadata!()).unwrap();
+    assert_eq!(HAS_INIT.load(Relaxed), true);
 }
