--- src/services/comp-sys/component/tests/kernel/Components.toml
@@ -1,5 +1,0 @@
-# template
-[components]
-kernel = {name = "kernel"}
-foo = {name = "foo"}
-bar = {name = "bar"}
